+++ pkgs/package_config/test/parse_test.dart
@@ -34,9 +34,9 @@ void main() {
 
       var foo = result["foo"];
       expect(foo, isNotNull);
-      expect(foo.root, Uri.parse("file:///foo/lib/"));
+      expect(foo.root, Uri.parse("file:///foo/"));
       expect(foo.packageUriRoot, Uri.parse("file:///foo/lib/"));
-      expect(foo.languageVersion, null);
+      expect(foo.languageVersion, LanguageVersion(2, 7));
     });
 
     test("valid empty", () {